working on data type issues. Everything had been strings now there are a few types of data used on csv

+++ src/main/java/org/launchcode/models/ISBN.java
@@ -5,7 +5,7 @@ package org.launchcode.models;
  */
 public class ISBN extends BookField {
 
-    public ISBN(Float value) {
+    public ISBN(int value) {
         super(value);
     }
 }
